Add recent review destinations section in Review Mode
A new feature is added to display recent review destinations in Review Mode. A scrollable box is implemented in the layout.pug file and styled in the custom.css. This box to be filled with the recent destinations fetched from the data during Review Mode initialization in the manage.js file. This addition helps users to quickly switch between frequently accessed reviews.